requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorzero==2.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
gpiozero==1.6.2
guizero==1.3.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
orjson==3.8.0
PyBoard==1.1.4
pydantic==1.10.2
pymata-aio==2.35
pyserial==3.5
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
RPi.GPIO==0.7.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.4.0
ujson==5.5.0
ULN2003Pi==1.0.4
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.18.0
websockets==10.3
wiringpi==2.60.1
